requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
psycopg2-binary==2.9.9
PyJWT==2.8.0
Flask-Bcrypt==1.0.1
Flask-APScheduler==1.13.1
python-dotenv==1.0.1
Flask-Cors==4.0.0
gunicorn==21.2.0
pillow==10.2.0
requests==2.31.0
